Simplify event store impl.
